Add systemd user service...
Same systemd service unit file was being used for system and user which works fine for
system service but for user service it causes a problem when a user enables the
unit. As install target is set to `multi-user.target` which doesn't
exist for user, it prints a warning and doesn't function as expected.

So instead create seperate user service unit, which changes its
install target to `default.target`.

Signed-off-by: Rahil Bhimjiani <me@rahil.rocks>